Move bank abstraction to infrastructure package
The infrastructure package includes the driven ports and its
implementations so that the application can execute its logic.
