+++ application/src/main/kotlin/funds_distribution.kt
@@ -1,6 +1,7 @@
 package lmirabal
 
 import lmirabal.finance.Amount
+import lmirabal.infrastructure.Bank
 import lmirabal.model.Account
 import lmirabal.model.AccountAddress
 import lmirabal.model.Pot
